ingestion-video/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
ffmpeg-python==0.2.0
numpy==1.24.3
opencv-python-headless==4.9.0.80
pillow==10.2.0
aiofiles==23.2.1
httpx==0.26.0
psycopg2-binary==2.9.9
minio==7.2.0
